+++ dcos/testdata/invalid/hour/state-summary.json
@@ -1,0 +1,2 @@
+---
+not a json
