+++ dcos/testdata/valid/hour/state-summary.json
@@ -1,0 +1,1 @@
+"{\"hostname\":\"172.17.0.2\",\"cluster\":\"DCOS\",\"slaves\":[{\"id\":\"f7b5b567-b49a-4b8c-8f71-acf46a23fdb9-S0\",\"hostname\":\"172.17.0.3\",\"port\":5051,\"attributes\":{},\"pid\":\"slave(1)@172.17.0.3:5051\",\"registered_time\":1538404905.241877,\"resources\":{\"disk\":20778.0,\"mem\":31012.0,\"gpus\":0.0,\"cpus\":8.0,\"ports\":\"[1025-2180, 2182-3887, 3889-5049, 5052-8079, 8082-8180, 8182-32000]\"},\"used_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"offered_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"reserved_resources\":{},\"unreserved_resources\":{\"disk\":20778.0,\"mem\":31012.0,\"gpus\":0.0,\"cpus\":8.0,\"ports\":\"[1025-2180, 2182-3887, 3889-5049, 5052-8079, 8082-8180, 8182-32000]\"},\"active\":true,\"version\":\"1.7.0\",\"capabilities\":[\"MULTI_ROLE\",\"HIERARCHICAL_ROLE\",\"RESERVATION_REFINEMENT\",\"RESOURCE_PROVIDER\",\"RESIZE_VOLUME\"],\"TASK_STAGING\":0,\"TASK_STARTING\":0,\"TASK_RUNNING\":0,\"TASK_KILLING\":0,\"TASK_FINISHED\":0,\"TASK_KILLED\":0,\"TASK_FAILED\":0,\"TASK_LOST\":0,\"TASK_ERROR\":0,\"TASK_UNREACHABLE\":0,\"framework_ids\":[]}],\"frameworks\":[{\"id\":\"f7b5b567-b49a-4b8c-8f71-acf46a23fdb9-0001\",\"name\":\"metronome\",\"pid\":\"scheduler-4191b30d-ab06-4940-92d2-5850a0bd20eb@172.17.0.2:15201\",\"used_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"offered_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"capabilities\":[\"PARTITION_AWARE\",\"REGION_AWARE\"],\"hostname\":\"172.17.0.2\",\"webui_url\":\"http://172.17.0.2:9000\",\"active\":true,\"connected\":true,\"recovered\":false,\"TASK_STAGING\":0,\"TASK_STARTING\":0,\"TASK_RUNNING\":0,\"TASK_KILLING\":0,\"TASK_FINISHED\":0,\"TASK_KILLED\":0,\"TASK_FAILED\":0,\"TASK_LOST\":0,\"TASK_ERROR\":0,\"TASK_UNREACHABLE\":0,\"slave_ids\":[]},{\"id\":\"f7b5b567-b49a-4b8c-8f71-acf46a23fdb9-0000\",\"name\":\"marathon\",\"pid\":\"scheduler-516a6289-68cf-416e-ae47-2c3e834b8f20@172.17.0.2:15101\",\"used_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"offered_resources\":{\"disk\":0.0,\"mem\":0.0,\"gpus\":0.0,\"cpus\":0.0},\"capabilities\":[\"TASK_KILLING_STATE\",\"GPU_RESOURCES\",\"PARTITION_AWARE\",\"REGION_AWARE\"],\"hostname\":\"172.17.0.2\",\"webui_url\":\"http://172.17.0.2:8080\",\"active\":true,\"connected\":true,\"recovered\":false,\"TASK_STAGING\":0,\"TASK_STARTING\":0,\"TASK_RUNNING\":0,\"TASK_KILLING\":0,\"TASK_FINISHED\":0,\"TASK_KILLED\":0,\"TASK_FAILED\":0,\"TASK_LOST\":0,\"TASK_ERROR\":0,\"TASK_UNREACHABLE\":0,\"slave_ids\":[]}]}"
