+++ units/helpers_windows.go
@@ -5,6 +5,7 @@ import (
 	"io"
 )
 
+// ReadJournalOutputSince returns error since windows does not support journal
 func ReadJournalOutputSince(unit, sinceString string) (io.ReadCloser, error) {
 	return nil, fmt.Errorf("there is no journal on Windows")
 }
